responses

+++ app/Http/Controllers/MarcaController.php
@@ -20,57 +20,71 @@ class MarcaController extends Controller
     public function index()
     {
         $marcas = $this->marca->all();
-        return $marcas;
+        return response()->json($marcas, 200);
     }
 
     /**
      * Store a newly created resource in storage.
      *
-     * @param  \Illuminate\Http\Request  $request
+     * @param \Illuminate\Http\Request $request
      * @return \Illuminate\Http\Response
      */
     public function store(Request $request)
     {
         $marca = $this->marca->create($request->all());
-        return $marca;
+        return response()->json($marca, 200);
     }
 
     /**
      * Display the specified resource.
      *
-     * @param  Integer
-     * @return \Illuminate\Http\Response
+     * @param Integer $id
+     * @return string[]
      */
     public function show($id)
     {
         $marca = $this->marca->find($id);
-        return $marca;
+        if ($marca === null) {
+            return response()->json([
+                'erro' => 'O recurso solicitado não existe.',
+            ], 404);
+        }
+        return response()->json($marca, 200);
     }
 
     /**
      * Update the specified resource in storage.
      *
-     * @param  \Illuminate\Http\Request  $request
-     * @param  Integer
+     * @param \Illuminate\Http\Request $request
+     * @param Integer $id
      * @return \Illuminate\Http\Response
      */
     public function update(Request $request, $id)
     {
         $marca = $this->marca->find($id);
+        if ($marca === null) {
+            $errormsg = ['erro' => 'Não foi possível atualizar os dados. O recurso solicitado não existe.'];
+            return response()->json($errormsg, 404);
+        }
         $marca->update($request->all());
-        return $marca;
+        return response()->json($marca, 200);
     }
 
     /**
      * Remove the specified resource from storage.
      *
-     * @param  Integer
+     * @param Integer
      * @return \Illuminate\Http\Response
      */
     public function destroy($id)
     {
         $marca = $this->marca->find($id);
+        if ($marca === null) {
+            $errormsg = ['erro' => 'Não foi possível realizar a exclusão. O recurso solicitado não existe.'];
+            return response()->json($errormsg, 404);
+        }
         $marca->delete();
-        return ['msg' => 'A Marca foi removida com sucesso.'];
+        $successmsg = ['msg' => 'A Marca foi removida com sucesso.'];
+        return response()->json($successmsg, 200);
     }
 }
